pyproject.toml
[project]
name = "scale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "colored>=2.3.1",
    "numpy-stl>=3.2.0",
    "pymadcad>=0.19.1",
    "yaspin>=3.3.0",
]
